bridge/requirements.txt
python-akismet
psycopg2-binary